AppEasePlatform/backendapi/requirements.txt
django-cors-headers
djangorestframework
